active reset pwd, fortify

+++ app/Http/Controllers/ScheduleController.php
@@ -4,19 +4,17 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Schedule;
+use Illuminate\Support\Facades\Auth;
 
 class ScheduleController extends Controller
 {
     public function index(Request $request)
     {
 
-        if($request->ajax()) {
-            $data = Schedule::whereDate('start', '>=',' $request->start')
-                ->whereDate('end',   '<=', $request->end)
-                ->get(['id', 'title', 'start', 'end']);
+        if ($request->ajax()) {
+            $data = Schedule::all();
 
             return response()->json($data);
-
         }
 
         return view('schedule');
@@ -27,39 +25,77 @@ class ScheduleController extends Controller
      *
      * @return response()
      */
-    public function ajax(Request $request)
+
+    public function action(Request $request)
     {
-
-        switch ($request->type) {
-            case 'add':
+    
+        if ($request->ajax()) {
+            if ($request->type == 'add') {
                 $event = Schedule::create([
-                    'title' => $request->title,
-                    'start' => $request->start,
-                    'end' => $request->end,
+                    'title'        =>    $request->title,
+                    'start'        =>    $request->start,
+                    'end'        =>    $request->end,
                 ]);
-
-                return response()->json($event);
-                break;
-
-            case 'update':
+            }
+            if ($request->type == 'addPrivate') {
+                $event = Schedule::create([
+                    'title'        =>    $request->title,
+                    'start'        =>    $request->start,
+                    'end'        =>    $request->end,
+                    'user_id' => Auth::user()->id
+                ]);
+            }
+            if ($request->type == 'update') {
                 $event = Schedule::find($request->id)->update([
-                    'title' => $request->title,
-                    'start' => $request->start,
-                    'end' => $request->end,
+                    'title'        =>    $request->title,
+                    'start'        =>    $request->start,
+                    'end'        =>    $request->end,
                 ]);
 
                 return response()->json($event);
-                break;
+            }
 
-            case 'delete':
+            if ($request->type == 'delete') {
                 $event = Schedule::find($request->id)->delete();
 
                 return response()->json($event);
-                break;
-
-            default:
-                # code...
-                break;
+            }
         }
     }
+
+    // public function ajax(Request $request)
+    // {
+
+    //     switch ($request->type) {
+    //         case 'add':
+    //             $event = Schedule::create([
+    //                 'title' => $request->title,
+    //                 'start' => $request->start,
+    //                 'end' => $request->end,
+    //             ]);
+
+    //             return response()->json($event);
+    //             break;
+
+    //         case 'update':
+    //             $event = Schedule::find($request->id)->update([
+    //                 'title' => $request->title,
+    //                 'start' => $request->start,
+    //                 'end' => $request->end,
+    //             ]);
+
+    //             return response()->json($event);
+    //             break;
+
+    //         case 'delete':
+    //             $event = Schedule::find($request->id)->delete();
+
+    //             return response()->json($event);
+    //             break;
+
+    //         default:
+    //             # code...
+    //             break;
+    //     }
+    // }
 }
